# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hangman App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /game/{id}:
    # our controller name
    x-swagger-router-controller: game
    get:
      operationId: getOne
      description: get a game
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a game
      # define the parameters
      parameters:
        - name: id
          description: Game id
          type: string
          in: path
          required: true
        - name: title
          description: Game properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Game"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delGame
      description: delete a game
      # define the parameters
      parameters:
        - name: id
          description: Game id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetGameResponse:
    required:
      - id
      - progress
      - guesses
    properties:
      id:
        type: string
      progress: 
        type: string
      guesses:
        type: number
  Game:
    type: object
    properties:
      progress:
        type: string
        description: task object name
      guesses:
        type: number
        description: task description
    required:
      - progress
      - guesses
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string